+++ index.d.ts
@@ -213,6 +213,7 @@ export type UseExpandedOptions<D extends object> = Partial<{
   getSubRows: (row: Row<D>, relativeIndex: number) => Array<Row<D>>
   manualExpandedKey: IdType<D>
   paginateExpandedRows: boolean
+  getResetExpandedDeps: (i: TableInstance) => Array<any>
 }>
 
 export interface UseExpandedHooks<D extends object> {
@@ -250,6 +251,7 @@ export type UseFiltersOptions<D extends object> = Partial<{
   manualFilters: boolean
   disableFilters: boolean
   filterTypes: Filters<D>
+  getResetFiltersDeps: (i: TableInstance) => Array<any>
 }>
 
 export interface UseFiltersState<D extends object> {
@@ -323,6 +325,7 @@ export type UseGroupByOptions<D extends object> = Partial<{
     rows: Array<Row<D>>,
     columnId: IdType<D>
   ) => Record<string, Row<D>>
+  getResetGroupByDeps: (i: TableInstance) => Array<any>
 }>
 
 export interface UseGroupByHooks<D extends object> {
@@ -403,7 +406,7 @@ export namespace usePagination {
 export type UsePaginationOptions<D extends object> = Partial<{
   pageCount: number
   manualPagination: boolean
-  disablePageResetOnDataChange: boolean
+  getResetPageDeps: (i: TableInstance) => Array<any>
   paginateExpandedRows: boolean
 }>
 
@@ -435,6 +438,7 @@ export namespace useRowSelect {
 
 export type UseRowSelectOptions<D extends object> = Partial<{
   manualRowSelectedKey: IdType<D>
+  getResetSelectedRowPathsDeps: (i: TableInstance) => Array<any>
 }>
 
 export interface UseRowSelectHooks<D extends object> {
@@ -525,6 +529,7 @@ export type UseSortByOptions<D extends object> = Partial<{
     directions: boolean[]
   ) => Array<Row<D>>
   sortTypes: Record<string, SortByFn<D>>
+  getResetSortByDeps: (i: TableInstance) => Array<any>
 }>
 
 export interface UseSortByHooks<D extends object> {